bro why

+++ audio.py
@@ -85,7 +85,7 @@ class AudioHandler:
                 else:
                     state = self.sys.state_history[header]  # type: ignore
                 dec_chunk = decrypt_audio(chunk, header, audio_nonce, auth_tag, state)  # type: ignore
-                audio_array = np.frombuffer(dec_chunk, dtype="int16")
+                audio_array = np.frombuffer(dec_chunk, dtype="<i2")
                 audio_array = np.reshape(audio_array, (-1, channels))  # type: ignore
                 print("dec audio", audio_array[:100])
                 stream.write(audio_array)
